+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.44.0] - 2024-08-31
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): Capability to define a time preparation for meal.
+
 ## [1.43.3]
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Refreshing access token every 1 minute.